Add files via upload
LEDControl and tm1638-library-master are the necessary Arduino libraries to control the LED Matrix and the TM1638 LED & Key board.  These are the starting point for the development.

I've already seen a few potential areas that I'd like to improve in LEDControl to make it easier to use and manipulate the LED Matrices.  For example, I intend to create a simple Conway's Life pattern generator and it would be convenient to be able to copy the LEDControl internal state matrix.  I'd also like to see if copying the whole of a matrix could speed up the display process (it gets a bit slow when there are six cascaded matrices!).  Matrix88 is a provisional library dedicated to handling cascaded 8x8 LED matrices.